Update package.json version, use drawer to delete a transaction apart from adding one
