+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "expense-tracker-nextjs",
-  "version": "1.1.4",
+  "version": "1.2.4",
   "private": true,
   "engines": {
     "node": ">=18.0.0",
